read all, read one by id mapping complete.

+++ src/main/java/com/careerdevs/vehiclesRESTpractice/Car.java
@@ -7,7 +7,7 @@ public class Car extends Vehicle {
 
     public Car() {}
 
-    public Car(Long id, String make, String model, Integer year, Long currentOdometer) {
+    public Car(Long id, String make, String model, Integer year, Integer currentOdometer) {
         super(id,make, model, year, currentOdometer);
 
     }
